# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API",
        "contact": {
            "name": "Support Team",
            "email": "support@my-api.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/empresa": {
            "get": {
                "tags": [
                    "Empresa"
                ],
                "summary": "Obtener todas las empresas",
                "description": "Retorna una lista de empresas",
                "operationId": "c483da1cdbf1bf1e929aefdb16ff6bfe",
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "nit": {
                                                "type": "string"
                                            },
                                            "razon_social": {
                                                "type": "string"
                                            },
                                            "representante_legal": {
                                                "type": "string"
                                            },
                                            "FechaCreacion": {
                                                "type": "string",
                                                "format": "date"
                                            },
                                            "IdMunicipioFk": {
                                                "type": "integer"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error del servidor"
                    }
                }
            },
            "post": {
                "tags": [
                    "Empresa"
                ],
                "summary": "Crear una nueva empresa",
                "description": "Crea una nueva empresa y la retorna",
                "operationId": "ac33ff7059381cf2a8bff6bdfd82d420",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nit": {
                                        "type": "string"
                                    },
                                    "razon_social": {
                                        "type": "string"
                                    },
                                    "representante_legal": {
                                        "type": "string"
                                    },
                                    "FechaCreacion": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "IdMunicipioFk": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Empresa creada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Datos de entrada no válidos"
                    },
                    "500": {
                        "description": "Error del servidor"
                    }
                }
            }
        },
        "/api/empresa/{id}": {
            "get": {
                "tags": [
                    "Empresa"
                ],
                "summary": "Obtener una empresa por ID",
                "description": "Retorna una empresa por su ID",
                "operationId": "50670ff5ea67095a34f8d61673014f4c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "nit": {
                                            "type": "string"
                                        },
                                        "razon_social": {
                                            "type": "string"
                                        },
                                        "representante_legal": {
                                            "type": "string"
                                        },
                                        "FechaCreacion": {
                                            "type": "string",
                                            "format": "date"
                                        },
                                        "IdMunicipioFk": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error del servidor"
                    }
                }
            },
            "put": {
                "tags": [
                    "Empresa"
                ],
                "summary": "Actualizar una empresa por ID",
                "description": "Actualiza una empresa por su ID",
                "operationId": "dcf9ba751f4496e315d6a6396a833d75",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nit": {
                                        "type": "string"
                                    },
                                    "razon_social": {
                                        "type": "string"
                                    },
                                    "representante_legal": {
                                        "type": "string"
                                    },
                                    "FechaCreacion": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "IdMunicipioFk": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Empresa actualizada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Datos de entrada no válidos"
                    },
                    "500": {
                        "description": "Error del servidor"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Empresa"
                ],
                "summary": "Eliminar una empresa por ID",
                "description": "Elimina una empresa por su ID",
                "operationId": "6f337542dff3dcbea39b332ed755b56c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Empresa eliminada correctamente"
                    },
                    "500": {
                        "description": "Error del servidor"
                    }
                }
            }
        },
...
    },
...
}